+++ seguridad/vw_usuario.py
@@ -2,6 +2,11 @@ from django.shortcuts import render
 from django.urls import reverse
 from django.http import HttpResponseRedirect
 from django.contrib.auth.models import User
+from django.db.models import ProtectedError
+from django.conf import settings
+from random import randint
+from os.path import isfile
+from os import remove
 
 from .forms import *
 from .models import *
@@ -43,6 +48,17 @@ def new( request ):
             for g in request.POST.getlist( 'groups' ):
                 obj.groups.add( g )
             obj.save()
+            upload_to = 'usuarios'
+            if "" != request.POST.get( 'token' ):
+                file = "{}{}/{}.imgx".format( settings.MEDIA_ROOT, upload_to, request.POST.get( 'token' ) )
+                if isfile( file ):
+                    f = open( file, "r" )
+                    if "r" == f.mode:
+                        img = f.read().strip()
+                        f.close()
+                        obj.fotografia = "{}/{}".format( upload_to, img )
+                        obj.save()
+                        remove( file )
             return HttpResponseRedirect( reverse( 'usuario_ver', kwargs = { 'pk' : obj.pk } ) )
     frm = RegUsuario( request.POST or None )
     return render( request, 'global/form.html', {
@@ -50,11 +66,23 @@ def new( request ):
         'footer' : True,
         'titulo' : 'Usuarios',
         'titulo_descripcion' : 'Nuevo',
-        'frm' : frm
+        'frm' : frm,
+        'uploader' : {
+            'url' : settings.UPLOADER_URL,
+            'site' : settings.UPLOADER_SITE,
+            'key' : settings.UPLOADER_KEY,
+            'onresponse' : '',
+            'type' : 'usuarios',
+            'excecute' : '',
+            'token' : randint( 1, 999999999 ),
+            'message' : "Archivo Cargado",
+        }
     } )
 
 @valida_acceso( [ 'usr.usuarios_usuario' ] )
 def see( request, pk ):
+    if not Usr.objects.filter( pk = pk ).exists():
+        return HttpResponseRedirect( reverse( 'seguridad_item_no_encontrado' ) )
     obj = Usr.objects.get( pk = pk )
     frm = RegUsuario( instance = obj )
     arbol = [ obj ]
@@ -82,6 +110,8 @@ def see( request, pk ):
 
 @valida_acceso( [ 'usr.actualizar_usuarios_usuario' ] )
 def update( request, pk ):
+    if not Usr.objects.filter( pk = pk ).exists():
+        return HttpResponseRedirect( reverse( 'seguridad_item_no_encontrado' ) )
     usr = Usr.objects.get( pk = pk )
     if 'POST' == request.method:
         frm = RegUsuario( instance = usr, data = request.POST, files = request.FILES  )
@@ -94,26 +124,62 @@ def update( request, pk ):
             for g in request.POST.getlist( 'groups' ):
                 obj.groups.add( g )
             obj.save()
+            upload_to = 'usuarios'
+            if "" != request.POST.get( 'token' ):
+                file = "{}{}/{}.imgx".format( settings.MEDIA_ROOT, upload_to, request.POST.get( 'token' ) )
+                if isfile( file ):
+                    f = open( file, "r" )
+                    if "r" == f.mode:
+                        img = f.read().strip()
+                        f.close()
+                        obj.fotografia = "{}/{}".format( upload_to, img )
+                        obj.save()
+                        remove( file )
             return HttpResponseRedirect( reverse( 'usuario_ver', kwargs = { 'pk' : obj.pk } ) )
         else:
             return render( request, 'global/form.html', {
                 'menu_main' : Usr.objects.filter( id = request.user.pk )[ 0 ].main_menu_struct(),
-                    'footer' : True,
-                    'titulo' : 'Usuarios',
-                    'titulo_descripcion' : usr,
-                    'frm' : frm
+                'footer' : True,
+                'titulo' : 'Usuarios',
+                'titulo_descripcion' : usr,
+                'frm' : frm,
+                'uploader' : {
+                    'url' : settings.UPLOADER_URL,
+                    'site' : settings.UPLOADER_SITE,
+                    'key' : settings.UPLOADER_KEY,
+                    'onresponse' : '',
+                    'type' : 'usuarios',
+                    'excecute' : '',
+                    'token' : randint( 1, 999999999 ),
+                    'message' : "Archivo Cargado",
+                }
             } )
     else:
         frm = RegUsuario( instance = usr )
         return render( request, 'global/form.html', {
             'menu_main' : Usr.objects.filter( id = request.user.pk )[ 0 ].main_menu_struct(),
-                'footer' : True,
-                'titulo' : 'Usuarios',
-                'titulo_descripcion' : usr,
-                'frm' : frm
+            'footer' : True,
+            'titulo' : 'Usuarios',
+            'titulo_descripcion' : usr,
+            'frm' : frm,
+            'uploader' : {
+                'url' : settings.UPLOADER_URL,
+                'site' : settings.UPLOADER_SITE,
+                'key' : settings.UPLOADER_KEY,
+                'onresponse' : '',
+                'type' : 'usuarios',
+                'excecute' : '',
+                'token' : randint( 1, 999999999 ),
+                'message' : "Archivo Cargado",
+            }
         } )
 
 @valida_acceso( [ 'usr.eliminar_usuarios_usuario' ] )
 def delete( request, pk ):
-    Usr.objects.get( pk = pk ).delete()
-    return HttpResponseRedirect( reverse( 'usuario_inicio' ) )
+    try:
+        if not Usr.objects.filter( pk = pk ).exists():
+            return HttpResponseRedirect( reverse( 'seguridad_item_no_encontrado' ) )
+        Usr.objects.get( pk = pk ).delete()
+        return HttpResponseRedirect( reverse( 'usuario_inicio' ) )
+    except ProtectedError:
+        return HttpResponseRedirect( reverse( 'seguridad_item_con_relaciones' ) )
